refactor: move methods from `core` to `stac`
